requirements.txt
appium-python-client~=4.1.0
selenium~=4.24.0
requests~=2.32.3
opencv-python~=4.10.0.84
numpy~=2.1.1
torch==2.0.1
torchvision==0.15.2
onnx==1.16.0
onnxruntime==1.15.1
pycocotools==2.0.7
PyYAML==6.0.1
scipy==1.13.0
onnxslim==0.1.31
onnxruntime-gpu==1.18.0
gradio==4.31.5
psutil==5.9.8
py-cpuinfo==9.0.0
huggingface-hub==0.23.2
safetensors==0.4.3